moved the todoArray over to the App.js file
partially built out Todo.js and TodoList.js enough that they do pass props through each other from state.
