+++ src/App.js
@@ -1,4 +1,23 @@
 import React from 'react';
+import TodoList from './components/TodoComponents/TodoList';
+
+const todoArray = [
+  {
+      task: 'Drop Drycleaning',
+      id: 1528817077286,
+      completed: false
+  },
+  {
+      task: 'Pack Suitcase',
+      id: 1528817077289,
+      completed: false
+  },
+  {
+      task: 'Book Hotel',
+      id: 1528817077285,
+      completed: false
+  },
+];
 
 class App extends React.Component {
   // you will need a place to store your state in this component.
@@ -8,7 +27,7 @@ class App extends React.Component {
   constructor() {
     super();
     this.state = {
-
+      todos: todoArray
     }
   }
 
@@ -16,6 +35,7 @@ class App extends React.Component {
     return (
       <div>
         <h2>Welcome to your Todo App! {Date.now()} </h2>
+        <h3><TodoList todos={this.state.todos} /></h3>
       </div>
     );
   }
