+++ src/components/TodoComponents/Todo.js
@@ -1,11 +1,9 @@
 import React from 'react';
 
 const Todo = props => {
-    const { task } = props.todoProp;
+    // const { task } = props.todoProp;
     return (
-        <div>
-            <h3>{task}</h3>
-        </div>
+        <div>{props.todo.task}</div>
     );
 };
