+++ src/components/TodoComponents/TodoList.js
@@ -1,20 +1,23 @@
 // your components will all go in this `component` directory.
 // feel free to change this component.js into TodoList.js
-const todoArray = [
-    {
-        task: 'Drop Drycleaning',
-        id: 1528817077286,
-        completed: false
-    },
-    {
-        task: 'Pack Suitcase',
-        id: 1528817077289,
-        completed: false
-    },
-    {
-        task: 'Book Hotel',
-        id: 1528817077285,
-        completed: false
-    },
-];
+import React from 'react';
+import Todo from './Todo';
 
+
+
+const TodoList = props => {
+    return (
+        <div>
+        {/* {props.task} */}
+        {props.todos.map(todo => (
+            <Todo todo={todo} />
+        ) ) }
+        </div>
+    );
+};
+
+TodoList.defaultProps = {
+    task: 'default task',
+  };
+
+export default TodoList;
